Create TEXT EDITOR PROJECT
A text editor is a type of computer program that edits plain text. Such programs are sometimes known as "notepad" software. Text editors are provided with operating systems and software development packages and can be used to change files such as configuration files, documentation files and programming language source code.